+++ coil-base/src/main/java/coil/fetch/HttpFetcher.kt
@@ -1,10 +1,12 @@
 package coil.fetch
 
+import android.net.Uri
 import android.webkit.MimeTypeMap
 import androidx.annotation.VisibleForTesting
 import coil.bitmap.BitmapPool
 import coil.decode.DataSource
 import coil.decode.Options
+import coil.map.Mapper
 import coil.network.HttpException
 import coil.size.Size
 import coil.util.await
@@ -15,17 +17,38 @@ import okhttp3.HttpUrl
 import okhttp3.Request
 import okhttp3.ResponseBody
 
-internal class HttpUrlFetcher(private val callFactory: Call.Factory) : Fetcher<HttpUrl> {
+internal class HttpUriFetcher(callFactory: Call.Factory) : HttpFetcher<Uri>(callFactory) {
+
+    override fun handles(data: Uri) = data.scheme == "http" || data.scheme == "https"
+
+    override fun key(data: Uri) = data.toString()
+
+    override fun Uri.toHttpUrl(): HttpUrl = HttpUrl.get(toString())
+}
+
+internal class HttpUrlFetcher(callFactory: Call.Factory) : HttpFetcher<HttpUrl>(callFactory) {
 
     override fun key(data: HttpUrl) = data.toString()
 
+    override fun HttpUrl.toHttpUrl(): HttpUrl = this
+}
+
+internal abstract class HttpFetcher<T : Any>(private val callFactory: Call.Factory) : Fetcher<T> {
+
+    /**
+     * Perform this conversion in a [Fetcher] instead of a [Mapper] so
+     * [HttpUriFetcher] can execute [HttpUrl.get] on a background thread.
+     */
+    abstract fun T.toHttpUrl(): HttpUrl
+
     override suspend fun fetch(
         pool: BitmapPool,
-        data: HttpUrl,
+        data: T,
         size: Size,
         options: Options
     ): FetchResult {
-        val request = Request.Builder().url(data).headers(options.headers)
+        val url = data.toHttpUrl()
+        val request = Request.Builder().url(url).headers(options.headers)
 
         val networkRead = options.networkCachePolicy.readEnabled
         val diskRead = options.diskCachePolicy.readEnabled
@@ -52,7 +75,7 @@ internal class HttpUrlFetcher(private val callFactory: Call.Factory) : Fetcher<H
 
         return SourceResult(
             source = body.source(),
-            mimeType = getMimeType(data, body),
+            mimeType = getMimeType(url, body),
             dataSource = if (response.cacheResponse() != null) DataSource.DISK else DataSource.NETWORK
         )
     }