+++ coil-base/src/main/java/coil/intercept/EngineInterceptor.kt
@@ -48,6 +48,7 @@ import coil.util.setValid
 import coil.util.toDrawable
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.ensureActive
+import kotlinx.coroutines.withContext
 import kotlin.coroutines.coroutineContext
 import kotlin.math.abs
 
@@ -75,9 +76,6 @@ internal class EngineInterceptor(
             val data = request.data
             val size = chain.size
             val eventListener = chain.eventListener
-
-            // Mark the input data's bitmap as ineligible for pooling.
-            invalidateData(data)
 
             // Perform any data mapping.
             eventListener.mapStart(request, data)
@@ -108,55 +106,42 @@ internal class EngineInterceptor(
                 )
             }
 
-            // Decrement the value from the memory cache if it was not used.
-            if (value != null) referenceCounter.decrement(value.bitmap)
+            // Fetch, decode, transform, and cache the image on a background dispatcher.
+            return withContext(request.dispatcher) {
+                // Mark the input data as ineligible for pooling (if necessary).
+                invalidateData(request.data)
 
-            // Fetch and decode the image.
-            val (drawable, isSampled, dataSource) = execute(mappedData, fetcher, request, chain.requestType, size, eventListener)
+                // Decrement the value from the memory cache if it was not used.
+                if (value != null) referenceCounter.decrement(value.bitmap)
 
-            // Mark the drawable's bitmap as eligible for pooling.
-            validateDrawable(drawable)
+                // Fetch and decode the image.
+                val (drawable, isSampled, dataSource) =
+                    execute(mappedData, fetcher, request, chain.requestType, size, eventListener)
 
-            // Cache the result in the memory cache.
-            val isCached = writeToMemoryCache(request, memoryCacheKey, drawable, isSampled)
+                // Mark the drawable's bitmap as eligible for pooling.
+                validateDrawable(drawable)
 
-            return SuccessResult(
-                drawable = drawable,
-                request = request,
-                metadata = Metadata(
-                    memoryCacheKey = memoryCacheKey.takeIf { isCached },
-                    isSampled = isSampled,
-                    dataSource = dataSource,
-                    isPlaceholderMemoryCacheKeyPresent = chain.cached != null
+                // Cache the result in the memory cache.
+                val isCached = writeToMemoryCache(request, memoryCacheKey, drawable, isSampled)
+
+                // Return the result.
+                SuccessResult(
+                    drawable = drawable,
+                    request = request,
+                    metadata = Metadata(
+                        memoryCacheKey = memoryCacheKey.takeIf { isCached },
+                        isSampled = isSampled,
+                        dataSource = dataSource,
+                        isPlaceholderMemoryCacheKeyPresent = chain.cached != null
+                    )
                 )
-            )
+            }
         } catch (throwable: Throwable) {
             if (throwable is CancellationException) {
                 throw throwable
             } else {
                 return requestService.errorResult(chain.request, throwable)
             }
-        }
-    }
-
-    /** Prevent pooling the input data's bitmap. */
-    @Suppress("USELESS_CAST")
-    private fun invalidateData(data: Any) {
-        when (data) {
-            is BitmapDrawable -> referenceCounter.setValid(data.bitmap as Bitmap?, false)
-            is Bitmap -> referenceCounter.setValid(data, false)
-        }
-    }
-
-    /** Allow pooling the successful drawable's bitmap. */
-    private fun validateDrawable(drawable: Drawable) {
-        val bitmap = (drawable as? BitmapDrawable)?.bitmap
-        if (bitmap != null) {
-            // Mark this bitmap as valid for pooling (if it has not already been made invalid).
-            referenceCounter.setValid(bitmap, true)
-
-            // Eagerly increment the bitmap's reference count to prevent it being pooled on another thread.
-            referenceCounter.increment(bitmap)
         }
     }
 
@@ -202,8 +187,7 @@ internal class EngineInterceptor(
     }
 
     /** Return true if [cacheValue]'s size satisfies the [request]. */
-    @VisibleForTesting
-    internal fun isSizeValid(
+    private fun isSizeValid(
         cacheKey: MemoryCache.Key?,
         cacheValue: RealMemoryCache.Value,
         request: ImageRequest,
@@ -267,9 +251,29 @@ internal class EngineInterceptor(
         return true
     }
 
+    /** Prevent pooling the input data's bitmap. */
+    @Suppress("USELESS_CAST")
+    private fun invalidateData(data: Any) {
+        when (data) {
+            is BitmapDrawable -> referenceCounter.setValid(data.bitmap as Bitmap?, false)
+            is Bitmap -> referenceCounter.setValid(data, false)
+        }
+    }
+
+    /** Allow pooling the successful drawable's bitmap. */
+    private fun validateDrawable(drawable: Drawable) {
+        val bitmap = (drawable as? BitmapDrawable)?.bitmap
+        if (bitmap != null) {
+            // Mark this bitmap as valid for pooling (if it has not already been made invalid).
+            referenceCounter.setValid(bitmap, true)
+
+            // Eagerly increment the bitmap's reference count to prevent it being pooled on another thread.
+            referenceCounter.increment(bitmap)
+        }
+    }
+
     /** Load the [data] as a [Drawable]. Apply any [Transformation]s. */
-    @VisibleForTesting
-    internal suspend inline fun execute(
+    private suspend inline fun execute(
         data: Any,
         fetcher: Fetcher<Any>,
         request: ImageRequest,
@@ -290,7 +294,9 @@ internal class EngineInterceptor(
                     coroutineContext.ensureActive()
 
                     // Find the relevant decoder.
-                    val isDiskOnlyPreload = type == REQUEST_TYPE_ENQUEUE && request.target == null && !request.memoryCachePolicy.writeEnabled
+                    val isDiskOnlyPreload = type == REQUEST_TYPE_ENQUEUE &&
+                        request.target == null &&
+                        !request.memoryCachePolicy.writeEnabled
                     val decoder = if (isDiskOnlyPreload) {
                         // Skip decoding the result if we are preloading the data and writing to the memory cache is
                         // disabled. Instead, we exhaust the source and return an empty result.