+++ coil-base/src/test/java/coil/fetch/HttpFetcherTest.kt
@@ -2,6 +2,7 @@ package coil.fetch
 
 import android.content.Context
 import android.webkit.MimeTypeMap
+import androidx.core.net.toUri
 import androidx.test.core.app.ApplicationProvider
 import coil.bitmap.BitmapPool
 import coil.size.PixelSize
@@ -32,14 +33,13 @@ import kotlin.test.assertTrue
 
 @RunWith(RobolectricTestRunner::class)
 @OptIn(ExperimentalCoroutinesApi::class)
-class HttpUrlFetcherTest {
+class HttpFetcherTest {
 
     private lateinit var context: Context
     private lateinit var mainDispatcher: TestCoroutineDispatcher
     private lateinit var server: MockWebServer
     private lateinit var callFactory: Call.Factory
     private lateinit var pool: BitmapPool
-    private lateinit var fetcher: HttpUrlFetcher
 
     @Before
     fun before() {
@@ -48,7 +48,6 @@ class HttpUrlFetcherTest {
         server = createMockWebServer(context, "normal.jpg")
         callFactory = OkHttpClient()
         pool = BitmapPool(0)
-        fetcher = HttpUrlFetcher(callFactory)
     }
 
     @After
@@ -59,12 +58,13 @@ class HttpUrlFetcherTest {
 
     @Test
     fun `basic network URL fetch`() {
+        val fetcher = HttpUrlFetcher(callFactory)
         val url = server.url("/normal.jpg")
         assertTrue(fetcher.handles(url))
         assertEquals(url.toString(), fetcher.key(url))
 
         val result = runBlocking {
-            fetcher.fetch(pool, url, PixelSize(100, 100), createOptions())
+            fetcher.fetch(pool, url, PixelSize(100, 100), createOptions(context))
         }
 
         assertTrue(result is SourceResult)
@@ -73,12 +73,13 @@ class HttpUrlFetcherTest {
 
     @Test
     fun `basic network URI fetch`() {
-        val uri = server.url("/normal.jpg")
+        val fetcher = HttpUriFetcher(callFactory)
+        val uri = server.url("/normal.jpg").toString().toUri()
         assertTrue(fetcher.handles(uri))
         assertEquals(uri.toString(), fetcher.key(uri))
 
         val result = runBlocking {
-            fetcher.fetch(pool, uri, PixelSize(100, 100), createOptions())
+            fetcher.fetch(pool, uri, PixelSize(100, 100), createOptions(context))
         }
 
         assertTrue(result is SourceResult)
@@ -87,6 +88,8 @@ class HttpUrlFetcherTest {
 
     @Test
     fun `mime type is parsed correctly from content type`() {
+        val fetcher = HttpUriFetcher(callFactory)
+
         // https://android.googlesource.com/platform/frameworks/base/+/61ae88e/core/java/android/webkit/MimeTypeMap.java#407
         Shadows.shadowOf(MimeTypeMap.getSingleton()).addExtensionMimeTypMapping("svg", "image/svg+xml")