+++ coil-svg/src/androidTest/java/coil/decode/SvgDecoderTest.kt
@@ -37,7 +37,7 @@ class SvgDecoderTest {
                 pool = pool,
                 source = source,
                 size = PixelSize(400, 250), // coil_logo.svg's intrinsic dimensions are 200x200.
-                options = createOptions(scale = Scale.FIT)
+                options = createOptions(context = context, scale = Scale.FIT)
             )
         }
 
@@ -56,7 +56,7 @@ class SvgDecoderTest {
                 pool = pool,
                 source = source,
                 size = PixelSize(326, 50),
-                options = createOptions(scale = Scale.FILL)
+                options = createOptions(context = context, scale = Scale.FILL)
             )
         }